sator-ting-client/requirements.txt
PySocks==1.7.1
stem==1.8.2
